[Ignacio Solar] - Implementado requerimiento de misFavoritos

+++ Entrega CRUD/src/main/java/DBD/services/JuegoService.java
@@ -1,5 +1,7 @@
 package DBD.services;
 import java.util.List;
+
+import com.fasterxml.jackson.databind.JsonNode;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,6 +12,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 import DBD.models.Juego;
 import DBD.repositories.JuegoRepository;
 @CrossOrigin
@@ -29,17 +33,32 @@ public class JuegoService {
     }
     @PostMapping("/Juego")
     @ResponseBody
-    public Juego crear(@RequestBody Juego Juego){
-        return JuegoRepository.crear(Juego);
+    public Juego crear(@RequestBody Juego Juego, HttpServletRequest request){
+        HttpSession session = request.getSession();
+        int id_tipo = (int) session.getAttribute("ID_tipo");
+        if (id_tipo == 2 || id_tipo == 3){
+            return JuegoRepository.crear(Juego);
+        }
+        return null;
     }
     @PutMapping("/Juego/{id}")
     @ResponseBody
-    public Juego update(@RequestBody Juego Juego, @PathVariable int id){
+    public Juego update(@RequestBody Juego Juego, @PathVariable int id, HttpServletRequest request){
         Juego.setIDJuego(id);
-        return JuegoRepository.update(Juego);
+        HttpSession session = request.getSession();
+        int id_tipo = (int) session.getAttribute("ID_tipo");
+        if (id_tipo == 2 || id_tipo == 3){
+            return JuegoRepository.update(Juego);
+        }
+        return null;
     }
     @DeleteMapping("/Juego/{id}")
-    public String delete(@PathVariable int id){
-        return JuegoRepository.delete(id);
+    public String delete(@PathVariable int id, HttpServletRequest request){
+        HttpSession session = request.getSession();
+        int id_tipo = (int) session.getAttribute("ID_tipo");
+        if (id_tipo == 2 || id_tipo == 3){
+            return JuegoRepository.delete(id);
+        }
+        return null;
     }
 }
